Cambio los double por int, no había leído que son números enteros en este ejercicio.

+++ TEMA1/Ejercicio11.java
@@ -3,11 +3,11 @@ public class Ejercicio11 {
     public static void main(String[] args) {
         Scanner in = new Scanner(System.in);
         System.out.println("Escriba el primer número: ");
-        double a = in.nextDouble();
+        int a = in.nextInt();
         System.out.println("Escriba el segundo número: ");
-        double b = in.nextDouble();
+        int b = in.nextInt();
         System.out.println("Escriba el tercer número: ");
-        double c = in.nextDouble();
+        int c = in.nextInt();
         if ((b==a+1) && (c==b+1)){
             System.out.println("Los números son consecutivos.");
         }
